somehow bot4 is follwing, sonar broke tho have not tested picking up cans

+++ include/SensorArray.h
@@ -3,7 +3,7 @@ class SensorArray {
     SensorArray();
     int LFSensor[5];
     int digitalArr[6];
-    int error;
+    double error;
     int calculateError();
     bool anyFrontSensorOn();
     // int threshold_array[5];
